requirements.txt
Flask
flask-cors
pymongo
python-dotenv
PyPDF2
pytesseract
Pillow
PyMuPDF
google-generativeai
